requirements.txt
flask
langchain-google-genai
langchain
python-dotenv
google-generativeai